+++ src/items/weapons/weapons.h
@@ -190,7 +190,7 @@ class Weapon : public Script
 		}
 
 		uint32_t getManaCost(const Player* player) const;
-		int32_t getHealthCost(const Player* player) const;
+		int64_t getHealthCost(const Player* player) const;
 		bool executeUseWeapon(Player* player, const LuaVariant& var) const;
 
 		uint16_t id = 0;
